requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==2.2
django-bootstrap4==21.1
django-cors-headers==3.10.0
django-filter==21.1
django-on-heroku==1.1.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Markdown==3.3.4
psycopg2-binary==2.9.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
